feat($eslintignore): Don't format on save if file is ignored by .eslintignore
Allows the user to specify if they want to keep a file from being formatted on save if the file is
ignored by their .eslintignore file. This will allow users to avoid having to duplicate their
project-specific ignore globs into their settings for prettier-eslint. We can probably assume that
if they don't want to lint the file, then they don't want to prettify it. However, they can turn
this option off if they want (default is on). This commit also adds a helper library called
`atom-linter` that is used by projects like linter-eslint. It has a convenient recursive file
finding method that is helpful for recursively trying to find the nearest .eslintignore file.

Resolves #13